ggplot2.utils Package

ggplot2.utils provides simple access to utility functions extending ggplot2.

Details

Currently all of the functions are imported from other extension packages:

ggpp: geom_table() and associated functions.

EnvStats: stat_n_text() and associated functions.

ggstats: stat_prop() and associated functions.

See also

Author

Maintainer: Daniel Sabanés Bové daniel.sabanes_bove@rconis.com

Authors:

Samer Mouksassi (wrote original Kaplan-Meier code)

Michael Sachs (wrote original Kaplan-Meier code)

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]